refactor(sdk): absorb createCrossChainPermission into toSession
SessionDefinition.crossChainPermits now accepts the builder input shape directly; the resolver moves to an internal cross-chain-permits module consumed by expandCrossChainPermit. Export the input as CrossChainPermissionInput (with FromLeg/ToLeg); validUntil/validAfter are Date-only.
